PHOENIX-138 Create Routing Master

+++ egov/egov-eis/src/main/java/org/egov/eis/repository/PositionMasterRepository.java
@@ -50,7 +50,7 @@ import org.springframework.stereotype.Repository;
  *
  */
 @Repository
-public interface PositionMasterRepository extends JpaRepository<Position, Integer>{
+public interface PositionMasterRepository extends JpaRepository<Position, Long>{
 	
 	Position findByName(String name);
 	List<Position> findByNameContainingIgnoreCase(String name);
